Restructured Mongo Schemas and Added Some Potential Input Field Changes
